requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
decorator==4.4.2
Django==4.1.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.4
Pillow==9.2.0
proglog==0.1.10
PyJWT==2.5.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.5
urllib3==1.26.12
